requirements.txt
google-generativeai
python-dotenv
PyMuPDF
markdown
markdownify
streamlit